requirements.txt
vosk
sounddevice
numpy
flask
flask-cors
piper-tts
ollama
anthropic
python-dotenv
langchain
langchain-ollama
langchain-anthropic
langchain-core
langchain-google-genai
playwright
pyTelegramBotAPI
python-frontmatter
pymupdf
Pillow
faster_whisper
langchain_community
ddgs
deepagents